store: add PeerStore tests
